Utilize mint instead of safeMint
Because safeMint yields an exploit in our auction if we don't provide
claimable functionality, I suggest utilizing mint over safeMint
and optionally verifying the existence of `onERC721Received` in the
recipient ABI on auction creation and bid preflight checks.

Unless we are supporting a specific workflow onchain that requires
checking for the existence of this hook, it seems like unnecessary cost
to look for it.