Attempt to Send unwrapped Eth to participants
One of the product goals I've discussed with owners is minimizing
participant exposure to Weth when possible.

this change attempts to transfer eth back to participants and then
falls back to WETH if the transfer fails. We want to minimize extra
steps for users when possible while still maintaining a secure flow
and this accomplishes that.