feat(auction): allow for configurable auction length
A duration parameter has been added to the `createReserveAuction`
function. In an effort to save some gas and reduce needed logic,
I've implemented the required minimum duration of 1 day as a base
duration and the `_duration` parameter as additional duration beyond
that base duration. The upper limit enforced on the contract is
`type(uint32).max` but endingTime is uint256 so it converts safely.

test plan: unit tests added and updated